+++ resources/views/partials/social.blade.php
@@ -1,15 +1,21 @@
 @php
   $links = $links ?? \App\mh_social_links();
   $labeled = ! empty($labeled);
+  $cards = ! empty($cards);
 @endphp
 @if ($links)
-  <ul class="soc-list{{ $labeled ? ' soc-list--labeled' : '' }}">
+  <ul class="soc-list{{ $labeled ? ' soc-list--labeled' : '' }}{{ $cards ? ' soc-list--cards' : '' }}">
     @foreach ($links as $s)
       <li>
         <a class="soc-link" href="{{ esc_url($s['url']) }}"@if (($s['key'] ?? '') !== 'rss') rel="me noopener" target="_blank"@else rel="me"@endif>
           {!! \App\mh_svg_icon($s['key']) !!}
           @if ($labeled)
-            <span>{{ $s['label'] }}</span>
+            <span class="soc-link__copy">
+              <span class="soc-link__label">{{ $s['label'] }}</span>
+              @if ($cards && ! empty($s['note']))
+                <span class="soc-link__note">{{ $s['note'] }}</span>
+              @endif
+            </span>
           @else
             <span class="visually-hidden">{{ $s['label'] }}</span>
           @endif